requirements.txt
streamlit
python-dotenv
requests
transformers
torch
sentencepiece
